apiGenresFetch add
