+++ src/templates/movie-grid.hbs
@@ -4,12 +4,9 @@
         <div class="movie-grid__poster-container">
             <img class="movie-grid__poster-img" data-id="{{this.id}}" src=" {{this.poster_path}}" alt="Poster" />
         </div>
-        <h2 class="movie-grid__title" data-id="{{this.id}}">{{this.original_title}}</h2>
+        <h2 class="movie-grid__title" data-id="{{this.id}}">{{this.title}}</h2>
         <p class=" movie-grid__info">
-            {{#each this.genre_ids}}
-            <span>{{this}}</span>
-            {{/each}}
-            | {{this.release_date}}
+            {{this.genre_ids}} | {{this.release_date}}
         </p>
     </a>
 </li>
